tools: tpm2_certify: extend to accept multiple keys
For a TPM operation where the first parameter is of a TPM2B type
one can resuest the parameter to be encrypted. In oreder to do this
we need to be able pass, the object to be certified and two extra
objects and the auth values, one for auth and one for parameter
encryiption/decryption. This change make it possible to pass
two -c obj and two -p options, where the first set of values will
be use for HMAC and the second set will be used for ecn/dec.
For example:

tpm2_certify -C obj.context -P 0x0011 -c hmac_key.context -p 0x00FF -c
enc_key.context -p 0xFFFF -g 0x00B -a <fileName> -s <fileName>

Signed-off-by: Tadeusz Struk <tadeusz.struk@intel.com>